refactor(lex-core): split escape.rs and anchoring.rs into submodules
